requirements.txt
pyyaml==6.0
requests==2.26.0
configargparse==1.5.3
XlsxWriter==3.1.0